Add structure factor benchmark
